remove idea and iml file

+++ mall-member/src/main/java/com/dunshan/mall/member/service/imp/MemberAttentionServiceImpl.java
@@ -45,7 +45,7 @@ public class MemberAttentionServiceImpl implements MemberAttentionService {
     @Override
     public int delete(Long brandId) {
         UmsMember member = memberService.getCurrentMember();
-        return memberBrandAttentionRepository.deleteByMemberIdAndBrandId(member.getId(),brandId);
+        return (int) memberBrandAttentionRepository.deleteByMemberIdAndBrandId(member.getId(), brandId);
     }
 
     @Override
