+++ mall-member/src/main/java/com/dunshan/mall/member/service/imp/MemberCollectionServiceImpl.java
@@ -41,7 +41,7 @@ public class MemberCollectionServiceImpl implements MemberCollectionService {
     @Override
     public int delete(Long productId) {
         UmsMember member = memberService.getCurrentMember();
-        return productCollectionRepository.deleteByMemberIdAndProductId(member.getId(), productId);
+        return (int)productCollectionRepository.deleteByMemberIdAndProductId(member.getId(), productId);
     }
 
     @Override
